login & logout coverage, split tests by view type in tests dir
