+++ src/Clean-Energy-PA-Site/UserProfile/templates/password_reset_from_profile.html
@@ -1,0 +1,23 @@
+{% extends 'GreenEnergySearch/base.html' %}
+
+{% block title %}
+  Change Password
+{% endblock %}
+
+{% load crispy_forms_tags %}
+
+{% block content %}
+  <div class="container">
+    <div class="row justify-content-center">
+      <div class="col-md-4">
+        <form method="POST" class="form-group">
+          {% csrf_token %}
+          <div class="mb-3">{{ form.old_password|as_crispy_field }}</div>
+          <div class="mb-3">{{ form.new_password1|as_crispy_field }}</div>
+          <div class="mb-3">{{ form.new_password2|as_crispy_field }}</div>
+          <button type="submit" class="btn btn-success">Change Password</button>
+        </form>
+      </div>
+    </div>
+  </div>
+{% endblock %}
