+++ src/Clean-Energy-PA-Site/UserProfile/tests/test_reset_password_from_profile_view.py
@@ -1,0 +1,6 @@
+from UserProfile.tests.test_user_profile_base_view import UserProfileBaseTest
+
+
+class ResetPasswordFromProfileTest(UserProfileBaseTest):
+    def test(self):
+        pass
