+++ src/Clean-Energy-PA-Site/UserRegistration/forms.py
@@ -3,7 +3,6 @@ from django.contrib.auth import login, authenticate
 from django.contrib.auth import get_user_model
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
-from django.core.exceptions import ValidationError
 from uszipcode import SearchEngine
 
 User = get_user_model()
