+++ src/Clean-Energy-PA-Site/UserRegistration/templates/user_logout.html
@@ -1,0 +1,5 @@
+{% extends 'GreenEnergySearch/base.html' %}
+
+{% block title %}
+  Logout
+{% endblock %}
