+++ src/Clean-Energy-PA-Site/UserRegistration/tests/test_logout_view.py
@@ -1,0 +1,21 @@
+# future
+
+# standard library
+
+# third-party
+
+# Django
+from django.urls import reverse
+
+# local Django
+from UserRegistration.tests.test_base_view import BaseTest
+
+
+class LogoutViewTest(BaseTest):
+    def test_access_logout(self):
+        """Test ID: Test validates the logout page is accessible"""
+        self._login_user()
+        response = self.client.get(reverse("user_logout"))
+        self.assertEqual(response.status_code, 302)
+        self.assertRedirects(response, "/")
+        self.assertFalse(response.wsgi_request.user.is_authenticated)
