+++ src/Clean-Energy-PA-Site/UserRegistration/urls.py
@@ -5,7 +5,7 @@ app_name = "UserRegistration"
 
 urlpatterns = [
     path("register/", views.register, name="register"),
-    path("login/", views.user_login, name="login"),
-    path("logout/", views.user_logout, name="logout"),
+    path("/registration/login/", views.user_login, name="login"),
+    path("user_logout/", views.user_logout, name="user_logout"),
     path("activate/<uidb64>/<token>", views.activate, name="activate"),
 ]
